Move session Builder into Session module directly
