+++ src/session/index.ts
@@ -3,5 +3,4 @@
  */
 
 export * from "./session";
-export * from "./builder";
 export * from "./query";
