+++ src/session/session.ts
@@ -12,6 +12,29 @@ import * as registries from "ethpm/registries";
 
 import { Query } from "./query";
 
+export class Builder<T extends config.Config> {
+  private connectors: config.Connectors<T>;
+
+  constructor (config_: config.RawConfig<T>) {
+    this.connectors = Object.assign(
+      {}, ...Object.keys(config_)
+        .map((service) => ({
+          [service]: config.load(config_[service])
+        }))
+    );
+  }
+
+  async connect (options: any = {}): Promise<Session<T>> {
+    const workspace = Object.assign({}, ...await Promise.all(
+      Object.keys(this.connectors).map( async (service) => ({
+        [service]: await this.connectors[service].connect(options)
+      }))
+    ));
+
+    return new Session(workspace);
+  }
+}
+
 export class Session<T extends Config> {
   private workspace: config.Workspace<T>;
